refactor : Rank관련 기능 RankResult로 이동
