+++ src/main/java/lotto/domain/Lotto.java
@@ -4,6 +4,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
+import lotto.settings.LottoSettings;
 
 public class Lotto {
