+++ src/main/java/lotto/view/OutputView.java
@@ -4,6 +4,7 @@ import java.util.List;
 import lotto.domain.Lotto;
 import lotto.domain.LottoRank;
 import lotto.domain.Rank;
+import lotto.dto.RankResult;
 
 public class OutputView {
 
@@ -14,19 +15,14 @@ public class OutputView {
         }
     }
 
-    public static void renderingWinningStatistics(List<Rank> result) {
+    public static void renderingWinningStatistics(RankResult result) {
         LottoRank[] values = LottoRank.values();
         for (LottoRank lottoRank : values) {
             int matchCount = lottoRank.count();
             int price = lottoRank.price();
-            int winningCount = (int) result.stream().filter(r -> r.equals(lottoRank)).count();
+            int winningCount = result.count(lottoRank);
             System.out.println(String.format("%d개 일치 (%d원)- %d개", matchCount, price, winningCount));
         }
-        double statistics = calcPrice(result) / (double) (result.size() * 1000);
-        System.out.printf("총 수익률은 %.2f입니다.\n", statistics);
-    }
-
-    private static int calcPrice(List<Rank> result) {
-        return result.stream().mapToInt(Rank::price).sum();
+        System.out.printf("총 수익률은 %.2f입니다.\n", result.calcRate());
     }
 }
